requirements_docker_pytorch14.txt
torch==1.4.0+cpu
torchvision==0.5.0
albumentations==0.4.3
opencv-python==4.1.1.26
pillow==6.1.0
git+https://github.com/BloodAxe/pytorch-toolbelt@038b93d00b10a89b6a89e07ba24cf7fb8f76b27a
shapely==1.6.4
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
pandas==0.24.2
numpy==1.16.4
torchnet
tqdm
jpeg4py==0.1.4
catalyst==19.12
-f https://download.pytorch.org/whl/cpu/torch_stable.html